Revisione gestione dettaglio categorie e alert su categorie / conti. Rivisti i colori e le logiche di wait tra i caricamenti dell'homepage. Il bottone indietro di Android ora torna sulla homepage se premuto quando si entra nel dettaglio categorie

+++ app/src/main/java/com/bloomtregua/rgb/database/accounts/AccountDao.kt
@@ -23,5 +23,6 @@ interface AccountDao {
     fun getAllAccounts(): Flow<List<AccountEntity>>
 
     @Query("SELECT * FROM accounts WHERE accountId = :accountId")
-    fun getAccountById(accountId: Int): Flow<AccountEntity>
+    fun getAccountById(accountId: Long): Flow<AccountEntity>
+
 }
